requirements.txt
asgiref==3.2.10
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
Django==2.2
django-cors-headers==3.5.0
djangorestframework==3.12.1
djongo==1.2.30
gunicorn==20.0.4
idna==2.10
pycodestyle==2.6.0
pymongo==3.11.0
pytz==2020.1
requests==2.24.0
sqlparse==0.2.4
toml==0.10.1
urllib3==1.25.10

